Fixing the simulator. Changing types to native C/C++ types (UINT8 - uint8_t) A bug I introduced with simulated connection fixed, the callback was missing. Use of extern variable in simulator - type changed from array to NSArray - had to put a check around the lookup. Should probably fix the invalid index also. Fixed some code layout.

+++ heft/HeftConnection.mm
@@ -200,7 +200,7 @@ enum eBufferConditions{
         if(nwritten <= 0)
             throw communication_exception();
 
-        LOG(@"%@", ::dump(@"HeftConnection::WriteData : ", data, nwritten));
+        LOG(@"%@", ::dump(@"HeftConnection::WriteData : ", data, (int) nwritten));
         
         len -= nwritten;
         data += nwritten;
